Version 1 commit
This version reads all documents within a directory, and then places
documents into groups based on if a certain percentage of their words
(currently 25%) match.  Groups are all seeded by a document, and so
there will be as many groups as documents, and the word matching will
only check against the seeded document.  Due to checking against a
percentage of one of the documents word set, the relation is not
necessarily symmetric.